Scatterplot: Steals & Defensive Rating
As seen, there is little to no linear relationship between a player's steals and defensive rating, as there are players with low steal counts that still have a better defensive rating than those some with high steal counts. Most notably, there is a prominent outlier at the end of the graph of this season's leader in steals, Dyson Daniels. Daniels accumulated over 200 steals this season, yet has a defensive rating that is slightly worse than the league average.